refactor: API 테스트 전용 버튼 분리 #52
